Move mosaic controls to a separate component
This should improve performance because the mosaic component is now only
loaded when 'Background' is clicked
