+++ src/client/pages/Geo/GeoMap/GeoMapMenuMosaic/MosaicControl/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MosaicControl'
